+++ src/services/OptimizedImages.php
@@ -21,6 +21,7 @@ use craft\base\Component;
 use craft\base\ElementInterface;
 use craft\base\Field;
 use craft\base\Volume;
+use craft\console\Application as ConsoleApplication;
 use craft\elements\Asset;
 use craft\errors\ImageException;
 use craft\errors\SiteNotFoundException;
@@ -28,6 +29,7 @@ use craft\helpers\Image;
 use craft\helpers\Json;
 use craft\models\AssetTransform;
 use craft\models\FieldLayout;
+
 use yii\base\InvalidConfigException;
 
 /**
@@ -73,8 +75,9 @@ class OptimizedImages extends Component
      * @param Asset          $asset
      * @param array          $variants
      * @param OptimizedImage $model
+     * @param boolean        $force
      */
-    public function populateOptimizedImageModel(Asset $asset, $variants, OptimizedImage $model)
+    public function populateOptimizedImageModel(Asset $asset, $variants, OptimizedImage $model, $force = false)
     {
         Craft::beginProfile('populateOptimizedImageModel', __METHOD__);
         $settings = ImageOptimize::$plugin->getSettings();
@@ -97,7 +100,29 @@ class OptimizedImages extends Component
                     && Image::canManipulateAsImage($asset->getExtension())
                     && $asset->height > 0) {
                     // Create the transform based on the variant
+                    /** @var AssetTransform $transform */
                     list($transform, $aspectRatio) = $this->getTransformFromVariant($asset, $variant, $retinaSize);
+                    // If they want to $force it, set `fileExists` = 0 in the transform index, then delete the transformed image
+                    if ($force) {
+                        $transforms = Craft::$app->getAssetTransforms();
+                        try {
+                            $index = $transforms->getTransformIndex($asset, $transform);
+                            $index->fileExists = 0;
+                            $transforms->storeTransformIndexData($index);
+                            $volume = $asset->getVolume();
+                            $transformPath = $asset->folderPath . $transforms->getTransformSubpath($asset, $index);
+                            try {
+                                $volume->deleteFile($transformPath);
+                            } catch (\Throwable $exception) {
+                            }
+                        } catch (\Throwable $e) {
+                            $msg = 'Failed to update transform: '.$e->getMessage();
+                            Craft::error($msg, __METHOD__);
+                            if (Craft::$app instanceof ConsoleApplication) {
+                                echo $msg . PHP_EOL;
+                            }
+                        }
+                    }
                     // Only create the image variant if it is not upscaled, or they are okay with it being up-scaled
                     if (($asset->width >= $transform->width && $asset->height >= $transform->height)
                         || $settings->allowUpScaledImageVariants
@@ -105,13 +130,18 @@ class OptimizedImages extends Component
                         $this->addVariantImageToModel($asset, $model, $transform, $variant, $aspectRatio);
                     }
                 } else {
-                    Craft::error(
-                        'Could not create transform for: '.$asset->title
+                    $msg = 'Could not create transform for: '.$asset->title
                         .' - Final format: '.$finalFormat
                         .' - Element extension: '.$asset->getExtension()
-                        .' - canManipulateAsImage: '.Image::canManipulateAsImage($asset->getExtension()),
+                        .' - canManipulateAsImage: '.Image::canManipulateAsImage($asset->getExtension())
+                        ;
+                    Craft::error(
+                        $msg,
                         __METHOD__
                     );
+                    if (Craft::$app instanceof ConsoleApplication) {
+                        echo $msg . PHP_EOL;
+                    }
                 }
             }
         }
@@ -195,11 +225,12 @@ class OptimizedImages extends Component
     /**
      * @param Field            $field
      * @param ElementInterface $asset
+     * @param boolean          $force
      *
      * @throws \yii\db\Exception
      * @throws InvalidConfigException
      */
-    public function updateOptimizedImageFieldData(Field $field, ElementInterface $asset)
+    public function updateOptimizedImageFieldData(Field $field, ElementInterface $asset, $force = false)
     {
         /** @var Asset $asset */
         if ($asset instanceof Asset && $field instanceof OptimizedImagesField) {
@@ -216,7 +247,8 @@ class OptimizedImages extends Component
                 $this->populateOptimizedImageModel(
                     $asset,
                     $field->variants,
-                    $model
+                    $model,
+                    $force
                 );
             }
             // Save our field data directly into the content table
@@ -239,15 +271,18 @@ class OptimizedImages extends Component
     /**
      * Re-save all of the assets in all of the volumes
      *
+     * @param int|null $fieldId only for this specific id
+     * @param boolean Should image variants be forced to be recreated?
+     *
      * @throws \yii\base\InvalidConfigException
      */
-    public function resaveAllVolumesAssets()
+    public function resaveAllVolumesAssets($fieldId = null, $force = false)
     {
         $volumes = Craft::$app->getVolumes()->getAllVolumes();
         foreach ($volumes as $volume) {
             if (is_subclass_of($volume, Volume::class)) {
                 /** @var Volume $volume */
-                $this->resaveVolumeAssets($volume);
+                $this->resaveVolumeAssets($volume, $fieldId, $force);
             }
         }
     }
@@ -257,11 +292,12 @@ class OptimizedImages extends Component
      * OptimizedImages field in the FieldLayout
      *
      * @param Volume $volume for this volume
-     * @param null $fieldId only for this specific id
+     * @param int|null $fieldId only for this specific id
+     * @param boolean Should image variants be forced to be recreated?
      *
      * @throws InvalidConfigException
      */
-    public function resaveVolumeAssets(Volume $volume, $fieldId = null)
+    public function resaveVolumeAssets(Volume $volume, $fieldId = null, $force = false)
     {
         $needToReSave = false;
         /** @var FieldLayout $fieldLayout */
@@ -296,6 +332,7 @@ class OptimizedImages extends Component
                     'enabledForSite' => false,
                 ],
                 'fieldId' => $fieldId,
+                'force' => $force,
             ]));
             Craft::debug(
                 Craft::t(
@@ -314,8 +351,9 @@ class OptimizedImages extends Component
      * Re-save an individual asset
      *
      * @param int $id
+     * @param boolean Should image variants be forced to be recreated?
      */
-    public function resaveAsset(int $id)
+    public function resaveAsset(int $id, $force = false)
     {
         $queue = Craft::$app->getQueue();
         $jobId = $queue->push(new ResaveOptimizedImages([
@@ -325,6 +363,7 @@ class OptimizedImages extends Component
                 'status' => null,
                 'enabledForSite' => false,
             ],
+            'force' => $force,
         ]));
         Craft::debug(
             Craft::t(